Add initial support for Amazon EC2.
Like the "iso" target, this target does not run anything.  It creates
a directory structure which the user should embed in a EC2 volume which
can then be used to created a EC2 AMI and launched.  Since there are so
many degrees of freedom to EC2, handling the details of how exactly to
launch the resulting unikernel  is simply out of the scope of rumprun.