Default to -std=c++11
We do this here instead of in the specs file since I couldn't figure
out how to make -std=c++11 apply only to C++ for the preprocessor
options line.  Fix if you know how.

The only bug is that "foo-gcc foo.cc" will not use C++11, but I'm
not sure we care about that.