Specify C++ language standard and includes using compiler specs
To address issue #43, the gcc driver knows (based on a list of suffixes)
if it is compiling C or C++. So, we can use this knowledge in the specs
to set the required C++ standard.

Note that -std=c++11 must be specified in both the cpp_options and
cc1plus spec for this to work.

This change also sets the cpp options to only add the C++ include paths
when compiling C++ programs.